BackEnd/requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
fonttools==4.54.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mistune==3.0.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
